Split search backend test into command + implementation
